+++ .config/webpack/production.js
@@ -18,12 +18,7 @@ module.exports.create = function create() {
     c.optimization = {
       minimize: true,
       minimizer: [
-        new TerserPlugin({
-          terserOptions: {
-            mangle: false,    // This option has to be disabled, otherwise GPU function doesn't work for min files.
-            compress: false,  // This option has to be disabled, otherwise GPU function doesn't work for min files.
-          }
-        })
+        new TerserPlugin()
       ]
     };