+++ src/HyperFormula.ts
@@ -4292,7 +4292,6 @@ export class HyperFormula implements TypedEmitter {
    * @category Instance
    */
   public destroy(): void {
-    this._evaluator.interpreter.destroyGpu()
     objectDestroy(this)
   }