+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2021-09-23
+
 ### Changed
 
 - Removed `gpu.js` from optional dependencies and marked config options `gpujs` and `gpuMode` as deprecated.